+++ src/main/java/cart/controller/CartController.java
@@ -1,42 +1,17 @@
 package cart.controller;
 
 import cart.dto.ProductRequest;
-import cart.entity.ProductEntity;
 import cart.service.AdminService;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
-
-@Controller
+@RestController
 public class CartController {
 
     private final AdminService adminService;
 
     public CartController(AdminService adminService) {
         this.adminService = adminService;
-    }
-
-    @GetMapping("/")
-    String allProducts(Model model) {
-        return userAllProducts(model);
-    }
-
-    @GetMapping("/products")
-    String userAllProducts(Model model) {
-        List<ProductEntity> productEntities = adminService.selectAllProducts();
-        model.addAttribute("products", productEntities);
-        return "index";
-    }
-
-    @GetMapping("/admin")
-    String adminAllProducts(Model model) {
-        List<ProductEntity> productEntities = adminService.selectAllProducts();
-        model.addAttribute("products", productEntities);
-
-        return "admin";
     }
 
     @PostMapping("/admin/product")
@@ -47,8 +22,6 @@ public class CartController {
 
     @PostMapping("/admin/product/{id}")
     public ResponseEntity<String> updateProduct(@RequestBody ProductRequest productRequest, @PathVariable int id) {
-//        System.out.println("id = " + id);
-//        System.out.println("productRequest = " + productRequest);
         adminService.updateProduct(productRequest, id);
         return ResponseEntity.ok().build();
     }
